fix chat input ref
